refactor: refactor plugin structure and update exports
