+++ packages/docs/src/plugin.ts
