+++ packages/slides/src/plugin.ts
